refactor: rename codes to keys
